practice-app/obj/requirements.txt
Django==2.0.3
djangorestframework==3.7.7
ibm-watson >= 3.0.3 
